Langchain/requirements.txt
langchain 
langchain-text-splitters
langchain-community
bs4
langchain[google-genai]
langchain-google-genai
langchain-milvus
langchain_text_splitters 
langchain_core
pypdf
fastapi[standard]
pymilvus
python-dotenv
uvicorn
